[changed] Move built components to lib directory
In the past this library has suffered from accidentally pushing this
repo directly to npm without the built components. There has been a
separate repo to which the built source has been pushed and later
deployed from. While the approach is feasible it prevents you from using
the `npm link` feature, for easy testing of changes in your own
application.  This should address concerns like #325, #218, and any
others like them that I have missed. In particular this is the fix for
an automated setup as mentioned in #292.

If you are currently using components directly from this library, then
you will need to change your reference path. For example
`require('react-bootstrap/Alert')` should now be `require('react-bootstrap/lib/Alert')`.
`require('react-bootstrap').Alert` will still work. This applies to both
CommonJS and AMD usages.
