+++ docs/src/ReactPlayground.js
@@ -2,45 +2,45 @@ var React = require('react');
 var classSet = require('react/lib/cx');
 var CodeMirror = global.CodeMirror;
 var JSXTransformer = global.JSXTransformer;
-var Accordion = require('../../cjs/Accordion');
-var Alert = require('../../cjs/Alert');
-var Badge = require('../../cjs/Badge');
-var Button = require('../../cjs/Button');
-var ButtonGroup = require('../../cjs/ButtonGroup');
-var ButtonToolbar = require('../../cjs/ButtonToolbar');
-var Carousel = require('../../cjs/Carousel');
-var CarouselItem = require('../../cjs/CarouselItem');
-var Col = require('../../cjs/Col');
-var DropdownButton = require('../../cjs/DropdownButton');
-var Glyphicon = require('../../cjs/Glyphicon');
-var Grid = require('../../cjs/Grid');
-var Input = require('../../cjs/Input');
-var Jumbotron = require('../../cjs/Jumbotron');
-var Label = require('../../cjs/Label');
-var ListGroup = require('../../cjs/ListGroup');
-var ListGroupItem = require('../../cjs/ListGroupItem');
-var Nav = require('../../cjs/Nav');
-var Navbar = require('../../cjs/Navbar');
-var NavItem = require('../../cjs/NavItem');
-var MenuItem = require('../../cjs/MenuItem');
-var Modal = require('../../cjs/Modal');
-var ModalTrigger = require('../../cjs/ModalTrigger');
-var OverlayTrigger = require('../../cjs/OverlayTrigger');
-var OverlayMixin = require('../../cjs/OverlayMixin');
-var PageHeader = require('../../cjs/PageHeader');
-var PageItem = require('../../cjs/PageItem');
-var Pager = require('../../cjs/Pager');
-var Panel = require('../../cjs/Panel');
-var PanelGroup = require('../../cjs/PanelGroup');
-var Popover = require('../../cjs/Popover');
-var ProgressBar = require('../../cjs/ProgressBar');
-var Row = require('../../cjs/Row');
-var SplitButton = require('../../cjs/SplitButton');
-var TabbedArea = require('../../cjs/TabbedArea');
-var Table = require('../../cjs/Table');
-var TabPane = require('../../cjs/TabPane');
-var Tooltip = require('../../cjs/Tooltip');
-var Well = require('../../cjs/Well');
+var Accordion = require('../../lib/Accordion');
+var Alert = require('../../lib/Alert');
+var Badge = require('../../lib/Badge');
+var Button = require('../../lib/Button');
+var ButtonGroup = require('../../lib/ButtonGroup');
+var ButtonToolbar = require('../../lib/ButtonToolbar');
+var Carousel = require('../../lib/Carousel');
+var CarouselItem = require('../../lib/CarouselItem');
+var Col = require('../../lib/Col');
+var DropdownButton = require('../../lib/DropdownButton');
+var Glyphicon = require('../../lib/Glyphicon');
+var Grid = require('../../lib/Grid');
+var Input = require('../../lib/Input');
+var Jumbotron = require('../../lib/Jumbotron');
+var Label = require('../../lib/Label');
+var ListGroup = require('../../lib/ListGroup');
+var ListGroupItem = require('../../lib/ListGroupItem');
+var Nav = require('../../lib/Nav');
+var Navbar = require('../../lib/Navbar');
+var NavItem = require('../../lib/NavItem');
+var MenuItem = require('../../lib/MenuItem');
+var Modal = require('../../lib/Modal');
+var ModalTrigger = require('../../lib/ModalTrigger');
+var OverlayTrigger = require('../../lib/OverlayTrigger');
+var OverlayMixin = require('../../lib/OverlayMixin');
+var PageHeader = require('../../lib/PageHeader');
+var PageItem = require('../../lib/PageItem');
+var Pager = require('../../lib/Pager');
+var Panel = require('../../lib/Panel');
+var PanelGroup = require('../../lib/PanelGroup');
+var Popover = require('../../lib/Popover');
+var ProgressBar = require('../../lib/ProgressBar');
+var Row = require('../../lib/Row');
+var SplitButton = require('../../lib/SplitButton');
+var TabbedArea = require('../../lib/TabbedArea');
+var Table = require('../../lib/Table');
+var TabPane = require('../../lib/TabPane');
+var Tooltip = require('../../lib/Tooltip');
+var Well = require('../../lib/Well');
 
 var IS_MOBILE = typeof navigator !== 'undefined' && (
   navigator.userAgent.match(/Android/i)
