+++ ie8/src.js
@@ -1,21 +1,21 @@
 /** @jsx React.DOM */
 var React = require('react');
-var Button = require('../cjs/Button');
-var DropdownButton = require('../cjs/DropdownButton');
-var MenuItem = require('../cjs/MenuItem');
-var Accordion = require('../cjs/Accordion');
-var Panel = require('../cjs/Panel');
-var ButtonToolbar = require('../cjs/ButtonToolbar');
-var OverlayTrigger = require('../cjs/OverlayTrigger');
-var Tooltip = require('../cjs/Tooltip');
-var Alert = require('../cjs/Alert');
-var TabbedArea = require('../cjs/TabbedArea');
-var TabPane = require('../cjs/TabPane');
-var Modal = require('../cjs/Modal');
-var OverlayMixin = require('../cjs/OverlayMixin');
-var Popover = require('../cjs/Popover');
-var Carousel = require('../cjs/Carousel');
-var CarouselItem = require('../cjs/CarouselItem');
+var Button = require('../lib/Button');
+var DropdownButton = require('../lib/DropdownButton');
+var MenuItem = require('../lib/MenuItem');
+var Accordion = require('../lib/Accordion');
+var Panel = require('../lib/Panel');
+var ButtonToolbar = require('../lib/ButtonToolbar');
+var OverlayTrigger = require('../lib/OverlayTrigger');
+var Tooltip = require('../lib/Tooltip');
+var Alert = require('../lib/Alert');
+var TabbedArea = require('../lib/TabbedArea');
+var TabPane = require('../lib/TabPane');
+var Modal = require('../lib/Modal');
+var OverlayMixin = require('../lib/OverlayMixin');
+var Popover = require('../lib/Popover');
+var Carousel = require('../lib/Carousel');
+var CarouselItem = require('../lib/CarouselItem');
 
 var dropdownInstance = (
   <DropdownButton title="Dropdown">
