+++ test/AlertSpec.jsx
@@ -2,7 +2,7 @@
 
 var React          = require('react');
 var ReactTestUtils = require('react/lib/ReactTestUtils');
-var Alert          = require('../cjs/Alert');
+var Alert          = require('../lib/Alert');
 
 describe('Alert', function () {
   it('Should output a alert with message', function () {
