+++ test/NavItemSpec.jsx
@@ -2,7 +2,7 @@
 
 var React          = require('react');
 var ReactTestUtils = require('react/lib/ReactTestUtils');
-var NavItem        = require('../cjs/NavItem');
+var NavItem        = require('../lib/NavItem');
 
 describe('NavItem', function () {
   it('Should add active class', function () {
