+++ tools/amd/index.js
@@ -4,53 +4,53 @@ define(function (require) {
   'use strict';
 
   return {
-    Accordion: require('./Accordion'),
-    Affix: require('./Affix'),
-    AffixMixin: require('./AffixMixin'),
-    Alert: require('./Alert'),
-    BootstrapMixin: require('./BootstrapMixin'),
-    Badge: require('./Badge'),
-    Button: require('./Button'),
-    ButtonGroup: require('./ButtonGroup'),
-    ButtonToolbar: require('./ButtonToolbar'),
-    Carousel: require('./Carousel'),
-    CarouselItem: require('./CarouselItem'),
-    Col: require('./Col'),
-    CollapsableMixin: require('./CollapsableMixin'),
-    DropdownButton: require('./DropdownButton'),
-    DropdownMenu: require('./DropdownMenu'),
-    DropdownStateMixin: require('./DropdownStateMixin'),
-    FadeMixin: require('./FadeMixin'),
-    Glyphicon: require('./Glyphicon'),
-    Grid: require('./Grid'),
-    Input: require('./Input'),
-    Interpolate: require('./Interpolate'),
-    Jumbotron: require('./Jumbotron'),
-    Label: require('./Label'),
-    ListGroup: require('./ListGroup'),
-    ListGroupItem: require('./ListGroupItem'),
-    MenuItem: require('./MenuItem'),
-    Modal: require('./Modal'),
-    Nav: require('./Nav'),
-    Navbar: require('./Navbar'),
-    NavItem: require('./NavItem'),
-    ModalTrigger: require('./ModalTrigger'),
-    OverlayTrigger: require('./OverlayTrigger'),
-    OverlayMixin: require('./OverlayMixin'),
-    PageHeader: require('./PageHeader'),
-    Panel: require('./Panel'),
-    PanelGroup: require('./PanelGroup'),
-    PageItem: require('./PageItem'),
-    Pager: require('./Pager'),
-    Popover: require('./Popover'),
-    ProgressBar: require('./ProgressBar'),
-    Row: require('./Row'),
-    SplitButton: require('./SplitButton'),
-    SubNav: require('./SubNav'),
-    TabbedArea: require('./TabbedArea'),
-    Table: require('./Table'),
-    TabPane: require('./TabPane'),
-    Tooltip: require('./Tooltip'),
-    Well: require('./Well')
+    Accordion: require('./lib/Accordion'),
+    Affix: require('./lib/Affix'),
+    AffixMixin: require('./lib/AffixMixin'),
+    Alert: require('./lib/Alert'),
+    BootstrapMixin: require('./lib/BootstrapMixin'),
+    Badge: require('./lib/Badge'),
+    Button: require('./lib/Button'),
+    ButtonGroup: require('./lib/ButtonGroup'),
+    ButtonToolbar: require('./lib/ButtonToolbar'),
+    Carousel: require('./lib/Carousel'),
+    CarouselItem: require('./lib/CarouselItem'),
+    Col: require('./lib/Col'),
+    CollapsableMixin: require('./lib/CollapsableMixin'),
+    DropdownButton: require('./lib/DropdownButton'),
+    DropdownMenu: require('./lib/DropdownMenu'),
+    DropdownStateMixin: require('./lib/DropdownStateMixin'),
+    FadeMixin: require('./lib/FadeMixin'),
+    Glyphicon: require('./lib/Glyphicon'),
+    Grid: require('./lib/Grid'),
+    Input: require('./lib/Input'),
+    Interpolate: require('./lib/Interpolate'),
+    Jumbotron: require('./lib/Jumbotron'),
+    Label: require('./lib/Label'),
+    ListGroup: require('./lib/ListGroup'),
+    ListGroupItem: require('./lib/ListGroupItem'),
+    MenuItem: require('./lib/MenuItem'),
+    Modal: require('./lib/Modal'),
+    Nav: require('./lib/Nav'),
+    Navbar: require('./lib/Navbar'),
+    NavItem: require('./lib/NavItem'),
+    ModalTrigger: require('./lib/ModalTrigger'),
+    OverlayTrigger: require('./lib/OverlayTrigger'),
+    OverlayMixin: require('./lib/OverlayMixin'),
+    PageHeader: require('./lib/PageHeader'),
+    Panel: require('./lib/Panel'),
+    PanelGroup: require('./lib/PanelGroup'),
+    PageItem: require('./lib/PageItem'),
+    Pager: require('./lib/Pager'),
+    Popover: require('./lib/Popover'),
+    ProgressBar: require('./lib/ProgressBar'),
+    Row: require('./lib/Row'),
+    SplitButton: require('./lib/SplitButton'),
+    SubNav: require('./lib/SubNav'),
+    TabbedArea: require('./lib/TabbedArea'),
+    Table: require('./lib/Table'),
+    TabPane: require('./lib/TabPane'),
+    Tooltip: require('./lib/Tooltip'),
+    Well: require('./lib/Well')
   };
 });
